Implement Initial Slot Processing (#175)
* Implement updateProposerRandaoLayer

* Rename updateProposerRandaoSkips to updateProposerRandaoLayer
* Add validator_registry and shard_committees_at_slots initialization to
the default constructor of BeaconState to enable slot processing action
to be taken when Artemis is ran, without giving an error due to null variables.
* Create a public getBeacon_proposer_index method to BeaconState to be
able to call the method from the StateTransition module.

NOTES:
- I haven't been able to test it yet, due to failure to think of what values
to initialize the validator_registry and shard_committees_at_slots with.
- If UpdateProposerRandaoLayer can become a method of BeaconState
instead of StateTransition, there wouldn't be a public method
implemented for getBeacon_proposer_index. It's private counterpart
get_beacon_proposer_index could be used.

* Remove initiation of variables in constructor

Remove initiation of variables such as shard committees and validator registries
in BeaconState constructor. This will result in a lot of null pointer errors being
thrown when running.

* Ran spotlessApply

* Remove public duplicate function

* Remove getBeacon_proposer_index
* Change get_beacon_proposer_index from private to public

* Refactor local variable name

* Test updateProposerRandaoLayer method

* Test updateProposerRandaoLayer method in State Transition. Create a
StateTransitionTest file in the process.

* Implement and test updateLatestRandaoMixes

* Refactor state transition tests

* Refactor PR from snake to camel case

* implement get_previous_block_root

* refactor to setPrevious_block_root

* Fix setPrevious_block_root

According to spec previous block root corresponds to slot - 1 (assuming state slot number is updated already)

* Implement updateRecentBlockHashes

* Implement updateBatched_block_roots method in BeaconState in order to
make state changes to BeaconState object inside the class instead of in state transition object(during state transition).
* Make use of a function called merkle_roots in
updateBatched_block_roots
* Create a stub for merkle_roots

* moon

* add updateRecentBlockHashes

* move slotProcesssing to SlotProcessingUtil

* test and cleanup

+++ artemis/src/main/java/tech/pegasys/artemis/state/util/AttestationUtil.java
@@ -17,6 +17,7 @@ import tech.pegasys.artemis.Constants;
 import tech.pegasys.artemis.datastructures.beaconchainstate.PendingAttestationRecord;
 import tech.pegasys.artemis.ethereum.core.Hash;
 import tech.pegasys.artemis.state.BeaconState;
+import tech.pegasys.artemis.util.uint.UInt64;
 
 import java.util.ArrayList;
 import java.util.Iterator;
@@ -57,11 +58,12 @@ public class AttestationUtil {
         return current_epoch_boundary_attestations;
     }
 
+
     //https://github.com/ethereum/eth2.0-specs/blob/master/specs/core/0_beacon-chain.md#get_block_root
     public static Hash get_block_root(BeaconState state, long slot) throws Exception {
         long slot_upper_bound = slot + state.getLatest_block_roots().size();
         if((state.getSlot() <= slot_upper_bound) || slot < state.getSlot())
-            return state.getLatest_block_roots().get(Math.toIntExact(slot % state.getLatest_block_roots().size()));
+            return state.getLatest_block_roots().get(UInt64.valueOf(slot % state.getLatest_block_roots().size()));
         throw new BlockValidationException("Desired block root not within the provided bounds");
     }
